+++ mrp_multi_level/wizards/mrp_inventory_procure.py
@@ -25,6 +25,7 @@ class MrpInventoryProcure(models.TransientModel):
             "warehouse_id": planned_order.mrp_area_id.warehouse_id.id,
             "location_id": planned_order.product_mrp_area_id.location_proc_id.id
             or planned_order.mrp_area_id.location_id.id,
+            "supply_method": planned_order.product_mrp_area_id.supply_method,
         }
 
     @api.model
@@ -46,21 +47,19 @@ class MrpInventoryProcure(models.TransientModel):
     @api.model
     def default_get(self, fields):
         res = super(MrpInventoryProcure, self).default_get(fields)
-        mrp_inventory_obj = self.env["mrp.inventory"]
-        mrp_inventory_ids = self.env.context["active_ids"] or []
+        active_ids = self.env.context["active_ids"] or []
         active_model = self.env.context["active_model"]
-        if not mrp_inventory_ids or "item_ids" not in fields:
+        if not active_ids or "item_ids" not in fields:
             return res
-
-        assert active_model == "mrp.inventory", "Bad context propagation"
-
-        items = item_obj = self.env["mrp.inventory.procure.item"]
-        for line in mrp_inventory_obj.browse(mrp_inventory_ids).mapped(
-            "planned_order_ids"
-        ):
-            if line.qty_released < line.mrp_qty:
-                items += item_obj.create(self._prepare_item(line))
-        res["item_ids"] = [(6, 0, items.ids)]
+        if active_model == "mrp.inventory":
+            items = item_obj = self.env["mrp.inventory.procure.item"]
+            mrp_inventory_obj = self.env["mrp.inventory"]
+            for line in mrp_inventory_obj.browse(active_ids).mapped(
+                "planned_order_ids"
+            ):
+                if line.qty_released < line.mrp_qty:
+                    items += item_obj.create(self._prepare_item(line))
+            res["item_ids"] = [(6, 0, items.ids)]
         return res
 
     def make_procurement(self):
@@ -78,8 +77,8 @@ class MrpInventoryProcure(models.TransientModel):
                     item.qty,
                     item.uom_id,
                     item.location_id,
-                    "MRP: " + str(self.env.user.login),  # name?
-                    "MRP: " + str(self.env.user.login),  # origin?
+                    "MRP: " + item.planned_order_id.name or str(self.env.user.login),
+                    "MRP: " + item.planned_order_id.origin or str(self.env.user.login),
                     item.mrp_inventory_id.company_id,
                     values,
                 )
@@ -108,7 +107,7 @@ class MrpInventoryProcureItem(models.TransientModel):
     )
     qty = fields.Float(string="Quantity")
     uom_id = fields.Many2one(string="Unit of Measure", comodel_name="uom.uom")
-    date_planned = fields.Date(string="Planned Date", required=False)
+    date_planned = fields.Date(string="Planned Date", required=True)
     mrp_inventory_id = fields.Many2one(
         string="Mrp Inventory", comodel_name="mrp.inventory"
     )
@@ -116,6 +115,18 @@ class MrpInventoryProcureItem(models.TransientModel):
     product_id = fields.Many2one(string="Product", comodel_name="product.product")
     warehouse_id = fields.Many2one(string="Warehouse", comodel_name="stock.warehouse")
     location_id = fields.Many2one(string="Location", comodel_name="stock.location")
+    supply_method = fields.Selection(
+        string="Supply Method",
+        selection=[
+            ("buy", "Buy"),
+            ("none", "Undefined"),
+            ("manufacture", "Produce"),
+            ("pull", "Pull From"),
+            ("push", "Push To"),
+            ("pull_push", "Pull & Push"),
+        ],
+        readonly=True,
+    )
 
     def _prepare_procurement_values(self, group=False):
         return {
@@ -123,8 +134,8 @@ class MrpInventoryProcureItem(models.TransientModel):
                 fields.Date.from_string(self.date_planned)
             ),
             "warehouse_id": self.warehouse_id,
-            # 'company_id': self.company_id, # TODO: consider company
             "group_id": group,
+            "planned_order_id": self.planned_order_id.id,
         }
 
     @api.onchange("uom_id")