+++ mrp_warehouse_calendar/__manifest__.py
@@ -6,11 +6,10 @@
     "summary": "Considers the warehouse calendars in manufacturing",
     "version": "13.0.1.0.0",
     "license": "LGPL-3",
-    "website": "https://github.com/manufacture",
+    "website": "https://github.com/OCA/manufacture",
     "author": "ForgeFlow, Odoo Community Association (OCA)",
     "category": "Manufacturing",
     "depends": ["mrp", "stock_warehouse_calendar"],
     "installable": True,
-    "development_status": "Production/Stable",
     "maintainers": ["JordiBForgeFlow"],
 }